# docs/api/openapi.json
{
  "components": {
    "schemas": {
      "AisTrack": {
        "properties": {
          "mmsi": {
            "title": "Mmsi",
            "type": "string"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/AisTrackPoint"
            },
            "title": "Points",
            "type": "array"
          },
          "vessel_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vessel Name"
          }
        },
        "required": [
          "mmsi",
          "points"
        ],
        "title": "AisTrack",
        "type": "object"
      },
...
      "AuditTrail": {
        "description": "The 214 \u2192 7 cascade behind a detection's suspect list.\n\nModelled here because it previously was not: `/detections/{id}/audit`\nhad no response_model, so this shape existed only in the fixture dict\nand in the console's hand-written mirror, with nothing connecting the\ntwo. Defensibility is the product \u2014 the one endpoint that explains our\nreasoning should not be the one endpoint with no schema.",
        "properties": {
          "detection_id": {
            "title": "Detection Id",
            "type": "string"
          },
          "final_candidate_count": {
            "title": "Final Candidate Count",
            "type": "integer"
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/AuditStage"
            },
            "title": "Stages",
            "type": "array"
          }
        },
        "required": [
          "detection_id",
          "stages",
          "final_candidate_count"
        ],
        "title": "AuditTrail",
        "type": "object"
      },
      "BaselineGapProfile": {
        "description": "The per-vessel baseline that SCORING_MODEL \u00a72.1(b) requires before\na gap can be judged suspicious at all.\n\nField names and units mirror db/schema/001_ais_positions.sql's\n`ais_baseline_profiles` table directly (median_gap_seconds,\np95_gap_seconds, sample_count) rather than inventing API-layer names\n\u2014 this is meant to become close to a pass-through read of that table,\nand giving it different field names/units here would be exactly the\nkind of two-versions-of-the-truth drift this project keeps warning\nagainst elsewhere.\n\n`has_sufficient_history` is the one field NOT stored in the DB \u2014 it\nis derived at the API layer from `sample_count` (see\nMIN_SAMPLES_FOR_BASELINE) as a convenience so the console doesn't\nneed to know the threshold itself. `False` is the signal the UI uses\nto mark the AIS-gap-anomaly factor low-confidence: a vessel with too\nlittle history is unknown, never silently treated as average.",
        "properties": {
          "has_sufficient_history": {
            "title": "Has Sufficient History",
            "type": "boolean"
          },
          "median_gap_seconds": {
            "title": "Median Gap Seconds",
            "type": "number"
          },
          "p95_gap_seconds": {
            "title": "P95 Gap Seconds",
            "type": "number"
          },
          "sample_count": {
            "title": "Sample Count",
            "type": "integer"
          },
          "stale": {
            "title": "Stale",
            "type": "boolean"
          }
        },
        "required": [
          "sample_count",
          "median_gap_seconds",
          "p95_gap_seconds",
          "stale",
          "has_sufficient_history"
        ],
        "title": "BaselineGapProfile",
        "type": "object"
      },
...
      "Detection": {
        "properties": {
          "attributes": {
            "$ref": "#/components/schemas/DetectionAttributes"
          },
          "classification": {
            "enum": [
              "oil",
              "look_alike"
            ],
            "title": "Classification",
            "type": "string"
          },
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "confidence_raw": {
            "title": "Confidence Raw",
            "type": "number"
          },
          "geometry": {
            "$ref": "#/components/schemas/GeoPolygon"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "penalties": {
            "items": {
              "$ref": "#/components/schemas/Penalty"
            },
            "title": "Penalties",
            "type": "array"
          },
          "scene_id": {
            "title": "Scene Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "scene_id",
          "geometry",
          "confidence",
          "confidence_raw",
          "classification",
          "attributes",
          "penalties"
        ],
        "title": "Detection",
        "type": "object"
      },
      "DetectionAttributes": {
        "properties": {
          "area_km2": {
            "title": "Area Km2",
            "type": "number"
          },
          "damping_ratio_db": {
            "title": "Damping Ratio Db",
            "type": "number"
          },
          "edge_sharpness": {
            "title": "Edge Sharpness",
            "type": "number"
          },
          "major_axis_bearing_deg": {
            "title": "Major Axis Bearing Deg",
            "type": "number"
          },
          "thickness_class": {
            "enum": [
              "sheen",
              "thick_film"
            ],
            "title": "Thickness Class",
            "type": "string"
          }
        },
        "required": [
          "area_km2",
          "major_axis_bearing_deg",
          "damping_ratio_db",
          "edge_sharpness",
          "thickness_class"
        ],
        "title": "DetectionAttributes",
        "type": "object"
      },
...
      "GeoPolygon": {
        "properties": {
          "coordinates": {
            "items": {
              "items": {
                "maxItems": 2,
                "minItems": 2,
                "prefixItems": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  }
                ],
                "type": "array"
              },
              "type": "array"
            },
            "title": "Coordinates",
            "type": "array"
          },
          "type": {
            "const": "Polygon",
            "default": "Polygon",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "coordinates"
        ],
        "title": "GeoPolygon",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "Page_AisTrack_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AisTrack"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "Page[AisTrack]",
        "type": "object"
      },
      "Page_Detection_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Detection"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "Page[Detection]",
        "type": "object"
      },
...
      "ProblemDetail": {
        "description": "RFC 7807. Served with media type application/problem+json \u2014 see\nrouters/errors.py, which is the only place that constructs these.",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status"
        ],
        "title": "ProblemDetail",
        "type": "object"
      },
...
      "Suspect": {
        "properties": {
          "calibrated": {
            "title": "Calibrated",
            "type": "boolean"
          },
          "data_quality": {
            "$ref": "#/components/schemas/DataQualitySummary"
          },
          "factors": {
            "items": {
              "$ref": "#/components/schemas/SuspectFactor"
            },
            "title": "Factors",
            "type": "array"
          },
          "imo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Imo"
          },
          "inferred_release_ci_minutes": {
            "title": "Inferred Release Ci Minutes",
            "type": "number"
          },
          "inferred_release_utc": {
            "title": "Inferred Release Utc",
            "type": "string"
          },
          "mmsi": {
            "title": "Mmsi",
            "type": "string"
          },
          "posterior": {
            "title": "Posterior",
            "type": "number"
          },
          "rank": {
            "title": "Rank",
            "type": "integer"
          },
          "slick_age_ci": {
            "maxItems": 2,
            "minItems": 2,
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "title": "Slick Age Ci",
            "type": "array"
          },
          "slick_age_hours": {
            "title": "Slick Age Hours",
            "type": "number"
          },
          "vessel_name": {
            "title": "Vessel Name",
            "type": "string"
          },
          "vessel_type": {
            "title": "Vessel Type",
            "type": "string"
          }
        },
        "required": [
          "rank",
          "mmsi",
          "vessel_name",
          "vessel_type",
          "posterior",
          "calibrated",
          "inferred_release_utc",
          "inferred_release_ci_minutes",
          "slick_age_hours",
          "slick_age_ci",
          "factors",
          "data_quality"
        ],
        "title": "Suspect",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VesselStatic": {
        "properties": {
          "baseline_gap_profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BaselineGapProfile"
              },
              {
                "type": "null"
              }
            ]
          },
          "imo": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Imo"
          },
          "mmsi": {
            "title": "Mmsi",
            "type": "string"
          },
          "vessel_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vessel Name"
          },
          "vessel_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vessel Type"
          }
        },
        "required": [
          "mmsi"
        ],
        "title": "VesselStatic",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "SAGARDRISHTI API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/ais/tracks": {
      "get": {
        "operationId": "list_tracks_api_v1_ais_tracks_get",
        "parameters": [
          {
            "description": "minlon,minlat,maxlon,maxlat",
            "in": "query",
            "name": "bbox",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "minlon,minlat,maxlon,maxlat",
              "title": "Bbox"
            }
          },
          {
            "in": "query",
            "name": "start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "in": "query",
            "name": "end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          },
          {
            "in": "query",
            "name": "mmsi",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Mmsi"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_AisTrack_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Tracks",
        "tags": [
          "ais"
        ]
      }
    },
    "/api/v1/ais/vessels/{mmsi}": {
      "get": {
        "operationId": "get_vessel_api_v1_ais_vessels__mmsi__get",
        "parameters": [
          {
            "in": "path",
            "name": "mmsi",
            "required": true,
            "schema": {
              "title": "Mmsi",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VesselStatic"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "No such resource. problem+json, with `detail` naming the id."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Vessel",
        "tags": [
          "ais"
        ]
      }
    },
    "/api/v1/detections": {
      "get": {
        "operationId": "list_detections_api_v1_detections_get",
        "parameters": [
          {
            "in": "query",
            "name": "scene_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scene Id"
            }
          },
          {
            "in": "query",
            "name": "min_confidence",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Confidence"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_Detection_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Detections",
        "tags": [
          "detections"
        ]
      }
    },
    "/api/v1/detections/{detection_id}": {
      "get": {
        "operationId": "get_detection_api_v1_detections__detection_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "detection_id",
            "required": true,
            "schema": {
              "title": "Detection Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Detection"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "No such resource. problem+json, with `detail` naming the id."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Detection",
        "tags": [
          "detections"
        ]
      }
    },
    "/api/v1/detections/{detection_id}/audit": {
      "get": {
        "operationId": "get_audit_api_v1_detections__detection_id__audit_get",
        "parameters": [
          {
            "in": "path",
            "name": "detection_id",
            "required": true,
            "schema": {
              "title": "Detection Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditTrail"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "No such resource. problem+json, with `detail` naming the id."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Audit",
        "tags": [
          "detections"
        ]
      }
    },
    "/api/v1/detections/{detection_id}/evidence": {
      "post": {
        "operationId": "request_evidence_api_v1_detections__detection_id__evidence_post",
        "parameters": [
          {
            "in": "path",
            "name": "detection_id",
            "required": true,
            "schema": {
              "title": "Detection Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "Accepted \u2014 a job id is returned; poll /jobs/{id}."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "No such resource. problem+json, with `detail` naming the id."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "The endpoint is in the contract but its pipeline is not built. An expected state today, not a failure \u2014 clients branch on it."
          }
        },
        "summary": "Request Evidence",
        "tags": [
          "detections"
        ]
      }
    },
    "/api/v1/detections/{detection_id}/forecast": {
      "get": {
        "operationId": "get_forecast_api_v1_detections__detection_id__forecast_get",
        "parameters": [
          {
            "in": "path",
            "name": "detection_id",
            "required": true,
            "schema": {
              "title": "Detection Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "hours",
            "required": false,
            "schema": {
              "default": 48,
              "title": "Hours",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "No such resource. problem+json, with `detail` naming the id."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "The endpoint is in the contract but its pipeline is not built. An expected state today, not a failure \u2014 clients branch on it."
          }
        },
        "summary": "Get Forecast",
        "tags": [
          "detections"
        ]
      }
    },
    "/api/v1/detections/{detection_id}/hindcast": {
      "get": {
        "operationId": "get_hindcast_api_v1_detections__detection_id__hindcast_get",
        "parameters": [
          {
            "in": "path",
            "name": "detection_id",
            "required": true,
            "schema": {
              "title": "Detection Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "No such resource. problem+json, with `detail` naming the id."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "The endpoint is in the contract but its pipeline is not built. An expected state today, not a failure \u2014 clients branch on it."
          }
        },
        "summary": "Get Hindcast",
        "tags": [
          "detections"
        ]
      }
    },
    "/api/v1/detections/{detection_id}/suspects": {
      "get": {
        "operationId": "get_suspects_api_v1_detections__detection_id__suspects_get",
        "parameters": [
          {
            "in": "path",
            "name": "detection_id",
            "required": true,
            "schema": {
              "title": "Detection Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Suspect"
                  },
                  "title": "Response Get Suspects Api V1 Detections  Detection Id  Suspects Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                }
              }
            },
            "description": "No such resource. problem+json, with `detail` naming the id."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Suspects",
        "tags": [
          "detections"
        ]
      }
    },
...
  }
}